Update README.md
When  I was a kid, about 12, I used to play and write basic programs on  ATARI XL 800. I loved it and enjoyed it so much. I saved the whole of it in my warehouse. So around 2006, I found a whole world of atari 8 bits computers with games and emulators. I took my old ATARI 1050 old driver, and I was able to copy the data from the atari diskettes  I kept from the '80s, and it was nice to remember with old stuff like atari dos2,dos3, player missiles, and more. T

The atari emulator uses ATR disk format, and I wrote this small project to handle those files. It supports atari xl800 dos2 and dos3 file systems. Today, as a professional embedded SW engineer, I found that very basic concepts of today's' CPUs are quite similar to the old ones like the atari 6005 CPUs' and that old filesystem has very basics familiar with today's modern file systems.